Redisgn the Timeout proposal solution
The create node take the responsiblity to generate block when proposal
broadcast valid block or be reject by other nodes. but it is a
centralization solution, now we redesign the timeout case with a purely
decentralization solution which the new generated min signature node
will propsoal block in this case, it is purely decentralization solution
and we don't need or depend create node anymore.

Todo:
Add the aging lock to the signature list to avoid dead txn occupy the map

Signed-off-by: Yanbo Li <dreamfly281@gmail.com>